requirements.txt
flask==3.0.0
PyGithub==2.1.1
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0
uvicorn==0.27.1
gunicorn==21.2.0
python-json-logger==2.0.7
structlog==24.1.0 